Increased variable size for status/skill points to remove the 65k cap (bugreport:1579). Added sql upgrade script to adjust the appropriate char table columns. This is an enhancement to r5762.
git-svn-id: https://svn.code.sf.net/p/rathena/svn/trunk@13541 54d463be-8e91-2dee-dedb-b68131a5f0ec

+++ src/common/mmo.h
@@ -270,7 +270,7 @@ struct mmo_charstatus {
 	int zeny;
 
 	short class_;
-	unsigned short status_point,skill_point;
+	unsigned int status_point,skill_point;
 	int hp,max_hp,sp,max_sp;
 	unsigned int option;
 	short manner;
